removed some console logs
